--- homeworks/sargis_vardanyan/ChatDemo/Client1/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
