--- homeworks/sargis_vardanyan/ChatDemo/Client1/src/main/java/Client1.java
@@ -1,20 +1,0 @@
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.net.Socket;
-import java.util.Scanner;
-
-public class Client1 {
-    public static void main(String[] args) throws IOException {
-        Socket connectionToServer = new Socket("127.0.0.1", 8800);
-        Scanner sc = new Scanner(System.in);
-        PrintWriter writer = new PrintWriter(connectionToServer.getOutputStream());
-        writer.print("GET / Client1 / HTTP/1.1\r\n");
-        writer.flush();
-
-        String currentLine;
-        while (!(currentLine = sc.nextLine()).equals("wq")) {
-            writer.print("Client1: " + currentLine + "\r\n");
-            writer.flush();
-        }
-    }
-}
