--- homeworks/sargis_vardanyan/ChatDemo/Client2/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
