--- homeworks/sargis_vardanyan/ChatDemo/Server/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
